Feature/ Iniciando implementação das rotas de Monster.
