+++ Tibia.Ciclopedia.Application/Tibia.Ciclopedia.Application.csproj
@@ -7,6 +7,12 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <Compile Remove="UseCases\Update\**" />
+    <EmbeddedResource Remove="UseCases\Update\**" />
+    <None Remove="UseCases\Update\**" />
+  </ItemGroup>
+
+  <ItemGroup>
     <PackageReference Include="AutoMapper" Version="13.0.1" />
     <PackageReference Include="MediatR" Version="12.4.0" />
     <PackageReference Include="Microsoft.Extensions.Http" Version="8.0.1" />
