--- Tibia.Ciclopedia.Application/UseCases/GetItem/GetAll/GetAllItemsInput.cs
@@ -1,15 +1,0 @@
-﻿using MediatR;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Tibia.Ciclopedia.Application.BaseOutput;
-using Tibia.Ciclopedia.Domain.Items;
-
-namespace Tibia.Ciclopedia.Application.UseCases.GetItem.GetAll
-{
-	public class GetAllItemInput : IRequest<Output<IEnumerable<Item>>>
-	{
-	}
-}
