+++ Tibia.Ciclopedia.Application/UseCases/ItemUC/CreateItem/ICreateItemUseCase.cs
@@ -1,0 +1,14 @@
+﻿using MediatR;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using Tibia.Ciclopedia.Application.BaseOutput;
+
+namespace Tibia.Ciclopedia.Application.UseCases.ItemUC.CreateItem
+{
+    public interface ICreateItemUseCase : IRequestHandler<CreateItemInput, Output<Guid>>
+    {
+    }
+}
