+++ Tibia.Ciclopedia.Application/UseCases/ItemUC/GetItem/GetAll/GetAllItemsInput.cs
@@ -1,0 +1,15 @@
+﻿using MediatR;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using Tibia.Ciclopedia.Application.BaseOutput;
+using Tibia.Ciclopedia.Domain.Items;
+
+namespace Tibia.Ciclopedia.Application.UseCases.ItemUC.GetItem.GetAll
+{
+    public class GetAllItemInput : IRequest<Output<IEnumerable<Item>>>
+    {
+    }
+}
