+++ Tibia.Ciclopedia.Application/UseCases/ItemUC/GetItem/GetByName/GetItemByNameInput.cs
@@ -1,0 +1,16 @@
+﻿using MediatR;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using Tibia.Ciclopedia.Application.BaseOutput;
+using Tibia.Ciclopedia.Domain.Items;
+
+namespace Tibia.Ciclopedia.Application.UseCases.ItemUC.GetItem.GetByName
+{
+    public class GetItemByNameInput : IRequest<Output<Item>>
+    {
+        public string Name { get; set; }
+    }
+}
