--- Tibia.Ciclopedia.Application/UseCases/Update/UpdateItem/IUpdateItemUseCase.cs
@@ -1,9 +1,0 @@
-﻿using MediatR;
-using Tibia.Ciclopedia.Application.BaseOutput;
-
-namespace Tibia.Ciclopedia.Application.UseCases.Update.UpdateItem
-{
-	public interface IUpdateItemUseCase : IRequestHandler<UpdateItemInput, Output<bool>>
-	{
-	}
-}
