+++ Tibia.Ciclopedia.Domain/Monsters/Enums/MonsterDifficulty.cs
@@ -1,0 +1,18 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Tibia.Ciclopedia.Domain.Monsters.Enums
+{
+	public enum MonsterDifficulty
+	{
+		Harmless = 0,
+		Trivial = 1,
+		Easy = 2,
+		Medium = 3,
+		Hard = 4,
+		Challenging = 5,
+	}
+}
