+++ Tibia.Ciclopedia.Infrastructure.MongoDb/Collection/Members/ItemMember/PeriodControlCollectionMember.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Tibia.Ciclopedia.Infrastructure.MongoDb.Collection.Members.ItemMember
+{
+    public class PeriodControlCollectionMember
+    {
+        public DateTime TimeChecked { get; set; }
+        public DateTime TimeCheckedExpire { get; set; }
+    }
+}
