+++ Tibia.Ciclopedia.Infrastructure.MongoDb/Collection/Members/ItemMember/SlotsInfoItemCollectionMember.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Tibia.Ciclopedia.Infrastructure.MongoDb.Collection.Members.ItemMember
+{
+    public class SlotsInfoItemCollectionMember
+    {
+        public bool HaveSlots { get; set; }
+        public int Quantity { get; set; }
+    }
+}
