--- Tibia.Ciclopedia.Infrastructure.MongoDb/Collection/Members/SlotsInfoItemCollectionMember.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Tibia.Ciclopedia.Infrastructure.MongoDb.Collection.Members
-{
-	public class SlotsInfoItemCollectionMember
-	{
-		public bool HaveSlots { get; set; }
-		public int Quantity { get; set; }
-	}
-}
